Check if we are connected to the database, accounting for a false negative on the database having been installed.

+++ drupal8-apache/7.0/usr/local/share/drupal8/development/install_database.sh
@@ -1,5 +1,9 @@
 #!/bin/bash
 
+function database_connected() {
+  mysql -h"$DATABASE_HOST" -u"$DATABASE_USER" -p"$DATABASE_PASSWORD" -e "SELECT 1;"
+}
+
 function database_exists() {
   mysql -h"$DATABASE_HOST" -u"$DATABASE_USER" -p"$DATABASE_PASSWORD" "$DATABASE_NAME" -e "SHOW TABLES; SELECT FOUND_ROWS() > 0;" | grep -q 1
 }
@@ -11,40 +15,46 @@ if [ -f tools/assets/development/drupaldb.sql.gz ]; then
   fi
 
   set +e
+  database_connected
+  DATABASE_CONNECTED=$?
+
   database_exists
   DATABASE_EXISTS=$?
   set -e
 
-  if [ "$DATABASE_EXISTS" -ne 0 ]; then
+  if [ "$DATABASE_CONNECTED" -eq 0 ] && [ "$DATABASE_EXISTS" -ne 0 ]; then
     echo 'Create drupal database'
     echo "CREATE DATABASE IF NOT EXISTS $DATABASE_NAME ; GRANT ALL ON $DATABASE_NAME.* TO $DATABASE_USER@'%' IDENTIFIED BY '$DATABASE_PASSWORD' ; FLUSH PRIVILEGES" |  mysql -uroot -p"$DATABASE_ROOT_PASSWORD" -h"$DATABASE_HOST"
 
     echo 'zcating the drupal database dump into the database'
     zcat tools/assets/development/drupaldb.sql.gz | mysql -h"$DATABASE_HOST" -uroot -p"$DATABASE_ROOT_PASSWORD" "$DATABASE_NAME" || exit 1
   fi
-fi
-
-set +e
-database_exists
-DATABASE_EXISTS=$?
-set -e
-
-if [ "$DATABASE_EXISTS" -ne 0 ]; then
-  SETTINGS_DIR="/app/docroot/sites/default"
-
-  chmod u+w "$SETTINGS_DIR" || true
-  mkdir -p "$SETTINGS_DIR/files/"
-
-  # Install a database if there isn't one yet
+else
   set +e
-  as_code_owner "drush sql-query 'SHOW TABLES;' | grep -v cache | grep -q ''" /app/docroot
-  HAS_CURRENT_TABLES=$?
-  set -e
-  if [ "$HAS_CURRENT_TABLES" -ne 0 ] && [ -n "$DRUPAL_INSTALL_PROFILE" ]; then
-    chown "$CODE_OWNER:$CODE_GROUP" "$SETTINGS_DIR/files/"
-    as_code_owner "echo 'y' | drush site-install $DRUPAL_INSTALL_PROFILE" "/app/docroot"
-    chown -R "$APP_USER:$APP_GROUP" "$SETTINGS_DIR/files/"
-  fi
+  database_connected
+  DATABASE_CONNECTED=$?
 
-  chmod a-w "$SETTINGS_DIR"
+  database_exists
+  DATABASE_EXISTS=$?
+  set -e
+
+  if [ "$DATABASE_CONNECTED" -eq 0 ] && [ "$DATABASE_EXISTS" -ne 0 ]; then
+    SETTINGS_DIR="/app/docroot/sites/default"
+
+    chmod u+w "$SETTINGS_DIR" || true
+    mkdir -p "$SETTINGS_DIR/files/"
+
+    # Install a database if there isn't one yet
+    set +e
+    as_code_owner "drush sql-query 'SHOW TABLES;' | grep -v cache | grep -q ''" /app/docroot
+    HAS_CURRENT_TABLES=$?
+    set -e
+    if [ "$HAS_CURRENT_TABLES" -ne 0 ] && [ -n "$DRUPAL_INSTALL_PROFILE" ]; then
+      chown "$CODE_OWNER:$CODE_GROUP" "$SETTINGS_DIR/files/"
+      as_code_owner "echo 'y' | drush site-install $DRUPAL_INSTALL_PROFILE" "/app/docroot"
+      chown -R "$APP_USER:$APP_GROUP" "$SETTINGS_DIR/files/"
+    fi
+
+    chmod a-w "$SETTINGS_DIR"
+  fi
 fi
